requirements.txt
langchain_core
python-dotenv
streamlit
langchain_community
faiss-cpu
groq
langchain-groq
langchainhub
sentence_transformers